[REFACT] move signup controller to a dedicated file
