+++ distribution_circuits_website_sale/controllers/__init__.py
@@ -1,1 +1,2 @@
 from . import main
+from . import AuthSignupHome
